# chore: #125 improve webpack bundling performance