--- a/.github/workflows/pull-request-ci.yml
+++ b/.github/workflows/pull-request-ci.yml
@@ -38,8 +38,8 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-yarn-
 
-      # Currently, actions/caches not support multiple paths https://github.com/actions/cache/issues/16
-      # Looks like they're working on it, let keep an eye on it
+      # Currently, actions/caches does not support multiple paths https://github.com/actions/cache/issues/16
+      # Looks like they're working on it, let's keep an eye on it
       - name: Get marketplace build cache
         uses: actions/cache@v1
         with: