--- a/.github/workflows/pull-request-ci.yml
+++ b/.github/workflows/pull-request-ci.yml
@@ -38,6 +38,12 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-yarn-
 
+      - name: Get jest cache
+        uses: actions/cache@v1
+        with:
+          path: scripts/jest/.jest-cache
+          key: ${{ runner.os }}-jest-cache
+
       # Currently, actions/caches does not support multiple paths https://github.com/actions/cache/issues/16
       # Looks like they're working on it, let's keep an eye on it
       - name: Get marketplace build cache