--- a/.github/workflows/release-develop.yml
+++ b/.github/workflows/release-develop.yml
@@ -37,6 +37,7 @@ jobs:
           key: ${{ runner.os }}-yarn-${{ hashFiles('yarn.lock') }}
           restore-keys: |
             ${{ runner.os }}-yarn-
+
       - name: Install Global Dependencies
         run: |
           yarn global add aws-cli serverless@1.59.3