--- a/.gitignore
+++ b/.gitignore
@@ -60,3 +60,6 @@ invite-developers.js
 
 # eslint cache
 .eslintcache
+
+# webpack cache
+.webpack-cache