--- a/.gitignore
+++ b/.gitignore
@@ -64,3 +64,6 @@ invite-developers.js
 # webpack cache
 .webpack-cache
 scripts/webpack/webpack-sass-prod/.temp/
+
+# jest cache
+scripts/jest/.jest-cache