--- a/.gitignore
+++ b/.gitignore
@@ -63,7 +63,7 @@ invite-developers.js
 
 # webpack cache
 .webpack-cache
-scripts/webpack/webpack-sass-prod/.temp/
+scripts/webpack/webpack-sass-prod/.temp/*
 
 # jest cache
 scripts/jest/.jest-cache