--- a/package.json
+++ b/package.json
@@ -15,7 +15,8 @@
   "scripts": {
     "build": "lerna run build:prod --parallel",
     "fetch-config": "yarn config-manager getSecret reapit-marketplace-app-config",
-    "lint": "eslint --cache --ext=jsx,ts,tsx packages --fix",
+    "lint": "lerna run lint --parallel --stream",
+    "lint:fix": "lerna run lint:fix --parallel --stream",
     "release:dev": "lerna run release:dev --parallel",
     "release:prod": "lerna run release:prod --parallel",
     "test": "lerna run test:ci --parallel",
@@ -32,8 +33,8 @@
       "yarn lint:cypress",
       "git add"
     ],
-    "*.{ts,tsx}": [
-      "yarn lint",
+    "packages/**/*.{ts,tsx,js}": [
+      "yarn lint:fix",
       "git add"
     ]
   },
@@ -150,9 +151,10 @@
     "babel-plugin-styled-components": "^1.10.6",
     "babel-plugin-tailwind-components": "^0.5.10",
     "babel-polyfill": "^6.26.0",
+    "cache-loader": "^4.1.0",
     "chromedriver": "^2.43.1",
     "compare-versions": "^3.5.1",
-    "concurrently": "^4.1.1",
+    "concurrently": "^5.1.0",
     "core-js": "^3.4.1",
     "cross-env": "^6.0.3",
     "css-loader": "^3.0.0",
@@ -175,6 +177,7 @@
     "fork-ts-checker-webpack-plugin": "^1.3.4",
     "graphql-import": "^0.7.1",
     "graphql-tag": "^2.10.1",
+    "hard-source-webpack-plugin": "^0.13.1",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^4.2.0",
     "isomorphic-fetch": "^2.2.1",
@@ -208,6 +211,7 @@
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.1",
     "sw2dts": "^2.6.2",
+    "thread-loader": "^2.1.3",
     "ts-jest": "^25.2.0",
     "ts-loader": "^6.2.1",
     "ts-node": "^8.3.0",