--- a/packages/aml-checklist/package.json
+++ b/packages/aml-checklist/package.json
@@ -18,7 +18,8 @@
     "test:dev": "cross-env TZ=UTC jest --watch --verbose",
     "release:dev": "node ../../scripts/release/release-dev.js aml-checklist reapit-aml-checklist-dev",
     "release:prod": "node ../../scripts/release/release-prod.js aml-checklist reapit-aml-checklist-prod",
-    "test:update-badges": "yarn test:ci && jest-coverage-badges --input src/tests/coverage/coverage-summary.json --output src/tests/badges"
+    "test:update-badges": "yarn test:ci && jest-coverage-badges --input src/tests/coverage/coverage-summary.json --output src/tests/badges",
+    "lint": "concurrently \"tsc --noEmit\" \"eslint --cache --ext=ts,tsx,js src\""
   },
   "devDependencies": {
     "@reapit/foundations-ts-definitions": "2020-02-13"