--- a/packages/cognito-auth/package.json
+++ b/packages/cognito-auth/package.json
@@ -26,7 +26,8 @@
     "release:prod": "node ../../scripts/release/release-npm.js @reapit/cognito-auth",
     "test:ci": "cross-env TZ=UTC jest --ci --colors --coverage --silent --forceExit",
     "test:dev": "cross-env TZ=UTC jest --watch --verbose",
-    "test:update-badges": "yarn test:ci && jest-coverage-badges --input src/tests/coverage/coverage-summary.json --output src/tests/badges"
+    "test:update-badges": "yarn test:ci && jest-coverage-badges --input src/tests/coverage/coverage-summary.json --output src/tests/badges",
+    "lint": "concurrently \"tsc --noEmit\" \"eslint --cache --ext=ts,tsx,js src\""
   },
   "dependencies": {
     "amazon-cognito-identity-js": "~3.0.13",