--- a/packages/cognito-auth/tsconfig.json
+++ b/packages/cognito-auth/tsconfig.json
@@ -9,15 +9,10 @@
       "@reapit/cognito-auth": ["../cognito-auth/src"],
       "@reapit/elements": ["../elements/src"]
     },
+    "declaration": true,
     "declarationDir": "dist"
   },
 
   "include": ["src"],
-  "exclude": [
-    "public",
-    "dist",
-    "src/scripts",
-    "node_modules",
-    "src/tests/coverage"
-  ]
+  "exclude": ["public", "dist", "src/scripts", "node_modules", "src/tests/coverage"]
 }