--- a/packages/cognito-custom-mail-lambda/package.json
+++ b/packages/cognito-custom-mail-lambda/package.json
@@ -12,7 +12,8 @@
     "build:prod": "rimraf dist && tsc --p tsconfig.json && cp -rf ./src/mailer/templates/ejs ./dist/mailer/templates/ejs",
     "release:dev": "cross-env CI=true serverless deploy --stage dev",
     "release:prod": "node ../../scripts/release/release-serverless.js cognito-custom-mail-lambda",
-    "test:update-badges": "yarn test:ci && jest-coverage-badges --input src/tests/coverage/coverage-summary.json --output src/tests/badges"
+    "test:update-badges": "yarn test:ci && jest-coverage-badges --input src/tests/coverage/coverage-summary.json --output src/tests/badges",
+    "lint": "concurrently \"tsc --noEmit\" \"eslint --cache --ext=ts,tsx,js src\""
   },
   "repository": {
     "type": "git",