--- a/packages/geo-diary/package.json
+++ b/packages/geo-diary/package.json
@@ -18,7 +18,8 @@
     "test:dev": "cross-env TZ=UTC jest --watch --verbose",
     "release:dev": "node ../../scripts/release/release-dev.js geo-diary reapit-geo-diary-dev",
     "release:prod": "node ../../scripts/release/release-prod.js geo-diary reapit-geo-diary-prod",
-    "test:update-badges": "yarn test:ci && jest-coverage-badges --input src/tests/coverage/coverage-summary.json --output src/tests/badges"
+    "test:update-badges": "yarn test:ci && jest-coverage-badges --input src/tests/coverage/coverage-summary.json --output src/tests/badges",
+    "lint": "concurrently \"tsc --noEmit\" \"eslint --cache --ext=ts,tsx,js src\""
   },
   "dependencies": {},
   "devDependencies": {