--- a/packages/graphql-server/package.json
+++ b/packages/graphql-server/package.json
@@ -15,7 +15,8 @@
     "test:ci": "cross-env TZ=UTC jest --ci --colors --coverage --silent --forceExit",
     "test:dev": "cross-env TZ=UTC jest --watch --verbose",
     "release:dev": "serverless deploy --stage dev",
-    "release:prod": "node ../../scripts/release/release-serverless.js graphql-server"
+    "release:prod": "node ../../scripts/release/release-serverless.js graphql-server",
+    "lint": "concurrently \"tsc --noEmit\" \"eslint --cache --ext=ts,tsx,js src\""
   },
   "dependencies": {
     "@reapit/cognito-auth": "2.0.19",