--- a/packages/graphql-server/tsconfig.json
+++ b/packages/graphql-server/tsconfig.json
@@ -14,6 +14,7 @@
     "paths": {
       "@/*": ["src/*"],
       "@reapit/elements": ["../elements/src"],
+      "@reapit/cognito-auth": ["../cognito-auth/src"],
       "logger": ["../../scripts/logger/sentry-logger.ts"]
     }
   },