--- a/packages/lifetime-legal/package.json
+++ b/packages/lifetime-legal/package.json
@@ -11,7 +11,7 @@
   "author": "Reapit Ltd",
   "main": "./src/index.ts",
   "scripts": {
-    "build:prod": "node '../../scripts/webpack/webpack-sass-prod/index.js'",
+    "build:prod": "node ../../scripts/webpack/webpack-sass-prod/index.js",
     "start:dev": "webpack-dev-server --hot --progress --color --mode development --config ../../scripts/webpack/webpack.sass.dev.js",
     "start:prod": "serve public/dist -s -l 8080",
     "test:ci": "cross-env TZ=UTC jest --ci --colors --coverage --silent --forceExit",