--- a/scripts/webpack/webpack.base.dev.js
+++ b/scripts/webpack/webpack.base.dev.js
@@ -8,6 +8,8 @@ const ResolveTSPathsToWebpackAlias = require('ts-paths-to-webpack-alias')
 const { PATHS } = require('./constants')
 const { getVersionTag } = require('../release/utils')
 const config = require(PATHS.config)
+const hashFiles = require('../utils/hash-files')
+const HardSourceWebpackPlugin = require('hard-source-webpack-plugin')
 
 module.exports = {
   context: process.cwd(),
@@ -56,12 +58,27 @@ module.exports = {
         windows: false,
       },
     }),
+    new HardSourceWebpackPlugin({
+      cacheDirectory: path.join(process.cwd(), '.webpack-cache', 'hard-source/[confighash]'),
+      environmentHash: {
+        root: path.join(__dirname, '../..'),
+        directories: [],
+        files: ['yarn.lock'],
+      },
+    }),
   ],
   module: {
     rules: [
       {
         test: /.tsx?$/,
         use: [
+          {
+            loader: 'cache-loader',
+            options: {
+              cacheDirectory: path.join(process.cwd(), '.webpack-cache', 'cache-loader'),
+              cacheIdentifier: hashFiles([path.join(__dirname, '../..', 'yarn.lock')]),
+            },
+          },
           {
             loader: 'ts-loader',
             options: { transpileOnly: true },