--- a/scripts/webpack/webpack.base.prod.js
+++ b/scripts/webpack/webpack.base.prod.js
@@ -109,7 +109,6 @@ module.exports = {
           {
             loader: 'babel-loader',
             options: {
-              cacheDirectory: true,
               presets: [
                 [
                   '@babel/preset-env',