--- a/scripts/webpack/webpack.base.prod.js
+++ b/scripts/webpack/webpack.base.prod.js
@@ -82,7 +82,7 @@ module.exports = {
     }),
     new HardSourceWebpackPlugin({
       // each package has its own .webpack-cache
-      cacheDirectory: path.join(process.cwd(), '.webpack-cache', 'hard-source/[confighash]'),
+      cacheDirectory: `${PATHS.cacheWebpackDir}/hard-source/[confighash]`,
       environmentHash: {
         root: path.join(__dirname, '../..'),
         directories: [],
@@ -100,7 +100,7 @@ module.exports = {
             loader: 'cache-loader',
             options: {
               // each package has its own .webpack-cache
-              cacheDirectory: path.join(process.cwd(), '.webpack-cache', 'cache-loader'),
+              cacheDirectory: `${PATHS.cacheWebpackDir}/cache-loader`,
               // use yarn.lock at the root of the monorepo as hash, relative to this file
               cacheIdentifier: hashFiles([path.join(__dirname, '../..', 'yarn.lock')]),
             },