@vuhuucuong 2020-02-26
@willmcvay Yeah okay, I will do check with the team after finish updating the pipeline, and do benchmarking too, thanks :)